image-blower-backend/requirements.txt
flask
torch==2.0.1
realesrgan
basicsr
facexlib
opencv-python
Pillow
